+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #Changelog
 All Notable changes to `Parser` will be documented in this file
-## Unreleased - 0.1.0-alpha.3
+## 0.1.0-alpha.3 [2016-06-02]
 
 ### Added
 - #21 Enable `cookies` by default